cuoiky_demo/requirements.txt
websockets>=11.0
bcrypt>=4.0.0
PyJWT>=2.8.0
Flask>=3.0.0
python-dotenv>=1.0.0
cryptography>=41.0.0
